% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/looplets.R
\name{ER_graph}
\alias{ER_graph}
\title{generates a random ER graph.}
\usage{
ER_graph(average.degree, num.nodes, rseed = 1)
}
\arguments{
\item{average.degree}{expected node degree. p = average.degree/num.nodes.}

\item{num.nodes}{number of nodes of the graph.}

\item{rseed}{a seed value for the random number generator.}
}
\value{
the edge list of the graph ensemble.
}
\description{
generates a random ER graph. In ER model node v and u are connected with a fixed probability p.
}
\examples{
er.edgelist <- ER_graph(3, 100)
g.stat <- graph_loops_stats(er.edgelist)
}

