% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/looplets.R
\name{graph_loops_stats}
\alias{graph_loops_stats}
\title{counts number of loops of length 3, 4, ..., max.loop.length.}
\usage{
graph_loops_stats(edge.list, max.loop.length = 10)
}
\arguments{
\item{edge.list}{an edge list of an undirected graph. If the graph is directed, i.e. v->u exits but u->v doesn't then it may return the wrong result without warning}

\item{max.loop.length}{an upper bound for the loop lengths}
}
\value{
the statistics of the loops including number of loops and number of edges in the counted loops. If edge e is a part of for instance two loops it is counted once.
}
\description{
Counts number of loops of length 3, 4, ..., max.loop.length. The results are inclusive. That is if an edge belongs to a loop of size k it is not consider for loop of size k+1.
}
\examples{
er.edgelist <- ER_graph(3, 100)
g.stat <- graph_loops_stats(er.edgelist)
}

